Allow search bar to search for label and extradata
Moved almost all of the logic from TransactionListItem.vue to useTransactionInfo composable
Also deleted some tx state that didn't exist in BTC
