+++ src/components/BtcTransactionListItem.vue
@@ -13,7 +13,7 @@
             <CircleSpinner/>
         </div>
         <div
-            v-else-if="state === TransactionState.EXPIRED || state === TransactionState.INVALIDATED"
+            v-else-if="state === TransactionState.INVALIDATED"
             class="invalid nq-red"
         >
             <CrossIcon/>
@@ -42,7 +42,6 @@
             <div class="time-and-message">
                 <span v-if="state === TransactionState.NEW">{{ $t('not sent') }}</span>
                 <span v-else-if="state === TransactionState.PENDING">{{ $t('pending') }}</span>
-                <span v-else-if="state === TransactionState.EXPIRED">{{ $t('expired') }}</span>
                 <span v-else-if="state === TransactionState.INVALIDATED">{{ $t('invalid') }}</span>
                 <span v-else-if="dateTime">{{ dateTime }}</span>
 
@@ -90,12 +89,8 @@ import {
 } from '@nimiq/vue-components';
 import { TransactionState } from '@nimiq/electrum-client';
 import { SwapAsset } from '@nimiq/fastspot-api';
-import { SettlementStatus } from '@nimiq/oasis-api';
-import { useBtcAddressStore } from '../stores/BtcAddress';
 import { useFiatStore } from '../stores/Fiat';
 import { Transaction } from '../stores/BtcTransactions';
-import { useBtcLabelsStore } from '../stores/BtcLabels';
-import { useAccountStore } from '../stores/Account';
 import { twoDigit } from '../lib/NumberFormatting';
 import Avatar from './Avatar.vue';
 import Amount from './Amount.vue';
@@ -105,10 +100,7 @@ import UsdcIcon from './icons/UsdcIcon.vue';
 import BankIcon from './icons/BankIcon.vue';
 import SwapSmallIcon from './icons/SwapSmallIcon.vue';
 import { FIAT_PRICE_UNAVAILABLE, BANK_ADDRESS } from '../lib/Constants';
-import { useSwapsStore } from '../stores/Swaps';
-import { useTransactionsStore } from '../stores/Transactions';
-import { useAddressStore } from '../stores/Address';
-import { useOasisPayoutStatusUpdater } from '../composables/useOasisPayoutStatusUpdater';
+import { useBtcTransactionInfo } from '../composables/useBtcTransactionInfo';
 import TransactionListOasisPayoutStatus from './TransactionListOasisPayoutStatus.vue';
 
 export default defineComponent({
@@ -121,186 +113,26 @@ export default defineComponent({
     setup(props, context) {
         const constants = { FIAT_PRICE_UNAVAILABLE, BANK_ADDRESS };
 
-        const {
-            state: btcAddresses$,
-            activeInternalAddresses,
-            activeExternalAddresses,
-        } = useBtcAddressStore();
-
-        const {
-            getRecipientLabel,
-            getSenderLabel,
-        } = useBtcLabelsStore();
-
         const state = computed(() => props.transaction.state);
 
-        const inputsSent = computed(() => props.transaction.inputs.filter((input) =>
-            input.address && (activeInternalAddresses.value.includes(input.address)
-                || activeExternalAddresses.value.includes(input.address)
-            ),
-        ));
+        const transaction = computed(() => props.transaction);
 
-        const isIncoming = computed(() => inputsSent.value.length === 0);
+        const {
+            data,
+            isCancelledSwap,
+            isIncoming,
+            peerAddresses,
+            peerLabel,
+            outputsReceived,
+            outputsSent,
+            swapData,
+            swapInfo,
+            swapTransaction,
+        } = useBtcTransactionInfo(transaction);
 
-        const outputsReceived = computed(() => {
-            if (!isIncoming.value) return [];
-
-            const receivedToExternal = props.transaction.outputs
-                .filter((output) => output.address && activeExternalAddresses.value.includes(output.address));
-
-            if (receivedToExternal.length > 0) return receivedToExternal;
-
-            return props.transaction.outputs
-                .filter((output) => output.address && activeInternalAddresses.value.includes(output.address));
-        });
         const amountReceived = computed(() => outputsReceived.value.reduce((sum, output) => sum + output.value, 0));
 
-        const outputsSent = computed(() => isIncoming.value
-            ? []
-            : props.transaction.outputs.filter((output) =>
-                !output.address || !activeInternalAddresses.value.includes(output.address)),
-        );
         const amountSent = computed(() => outputsSent.value.reduce((sum, output) => sum + output.value, 0));
-
-        const { getSwapByTransactionHash } = useSwapsStore();
-        const swapInfo = computed(() => getSwapByTransactionHash.value(props.transaction.transactionHash));
-        const swapData = computed(() => (isIncoming.value ? swapInfo.value?.in : swapInfo.value?.out) || null);
-        useOasisPayoutStatusUpdater(swapData);
-        const isCancelledSwap = computed(() =>
-            swapInfo.value?.in && swapInfo.value?.out && swapInfo.value.in.asset === swapInfo.value.out.asset);
-
-        const swapTransaction = computed(() => {
-            if (!swapData.value) return null;
-
-            if (swapData.value.asset === SwapAsset.NIM) {
-                let swapTx = useTransactionsStore().state.transactions[swapData.value.transactionHash];
-                if (swapTx?.relatedTransactionHash) {
-                    // Avoid showing the swap proxy, instead show our related address. Note that we don't test for
-                    // the swap proxy detection extra data here as the swap tx holds htlc data instead. Only the related
-                    // tx holds the proxy identifying extra data.
-                    swapTx = useTransactionsStore().state.transactions[swapTx.relatedTransactionHash];
-                }
-                return swapTx || null;
-            }
-
-            return null;
-        });
-
-        // Data
-        const data = computed(() => {
-            if (swapData.value) {
-                if (!isCancelledSwap.value) {
-                    const message = context.root.$t('Sent {fromAsset} – Received {toAsset}', {
-                        fromAsset: isIncoming.value ? swapData.value.asset : SwapAsset.BTC,
-                        toAsset: isIncoming.value ? SwapAsset.BTC : swapData.value.asset,
-                    }) as string;
-
-                    // The TransactionListOasisPayoutStatus takes care of the second half of the message
-                    if (
-                        swapData.value.asset === SwapAsset.EUR
-                        && swapData.value.htlc?.settlement
-                        && swapData.value.htlc.settlement.status !== SettlementStatus.CONFIRMED
-                    ) return `${message.split('–')[0]} –`;
-
-                    return message;
-                }
-
-                return isIncoming.value ? context.root.$t('HTLC Refund') : context.root.$t('HTLC Creation');
-            }
-
-            // if ('hashRoot' in props.transaction.data) {
-            //     return context.root.$t('HTLC Creation');
-            // }
-            // if ('creator' in props.transaction.proof) {
-            //     return context.root.$t('HTLC Refund');
-            // }
-            // if ('hashRoot' in props.transaction.proof) {
-            //     return context.root.$t('HTLC Settlement');
-            // }
-
-            return '';
-        });
-
-        // Peer
-        const peerAddresses = computed(() => {
-            if (swapData.value) {
-                if (swapData.value.asset === SwapAsset.NIM && swapTransaction.value) {
-                    const swapPeerAddress = isIncoming.value
-                        ? swapTransaction.value.sender
-                        : swapTransaction.value.recipient;
-                    if (!useAddressStore().state.addressInfos[swapPeerAddress] // not one of our addresses -> proxy
-                        && !swapTransaction.value.relatedTransactionHash) {
-                        // Avoid displaying proxy address identicon until we know related address.
-                        return [''];
-                    }
-                    return [swapPeerAddress];
-                }
-                if (swapData.value.asset === SwapAsset.EUR) return [constants.BANK_ADDRESS];
-            }
-
-            return (isIncoming.value
-                ? props.transaction.inputs.map((input) => input.address || input.script)
-                : outputsSent.value.map((output) => output.address || output.script)
-            ).filter((address, index, array) => array.indexOf(address) === index); // dedupe
-        });
-        const peerLabel = computed(() => {
-            if (isCancelledSwap.value) {
-                return context.root.$t('Cancelled Swap') as string;
-            }
-
-            if (swapData.value) {
-                if (swapData.value.asset === SwapAsset.NIM && swapTransaction.value) {
-                    return useAddressStore().state.addressInfos[peerAddresses.value[0]]?.label
-                        // avoid displaying proxy address until we know related peer address
-                        || context.root.$t('Swap') as string;
-                }
-
-                if (swapData.value.asset === SwapAsset.USDC) {
-                    return context.root.$t('USD Coin') as string;
-                }
-
-                if (swapData.value.asset === SwapAsset.EUR) {
-                    return swapData.value.bankLabel || context.root.$t('Bank Account') as string;
-                }
-
-                return swapData.value.asset.toUpperCase();
-            }
-
-            if (isIncoming.value) {
-                // Search sender labels
-                const ownAddresses = outputsReceived.value.map((output) => output.address);
-                for (const address of ownAddresses) {
-                    if (!address) continue;
-                    const label = getSenderLabel.value(address);
-                    if (label) return label;
-                }
-            } else {
-                // Search recipient labels
-                for (const address of peerAddresses.value) {
-                    const label = getRecipientLabel.value(address);
-                    if (label) return label;
-                }
-            }
-
-            // Search other stored addresses
-            for (const address of peerAddresses.value) {
-                const ownedAddressInfo = btcAddresses$.addressInfos[address];
-                if (ownedAddressInfo) {
-                    // Find account label
-                    const { accountInfos } = useAccountStore();
-                    return Object.values(accountInfos.value)
-                        .find((accountInfo) => accountInfo.btcAddresses.external.includes(address))?.label
-                        || Object.values(accountInfos.value)
-                            .find((accountInfo) => accountInfo.btcAddresses.internal.includes(address))!.label;
-                }
-            }
-
-            // TODO: Search global address book
-            // const globalLabel = AddressBook.getLabel(peerAddress.value);
-            // if (globalLabel) return globalLabel;
-
-            return false;
-        });
 
         // Date
         const date = computed(() => props.transaction.timestamp && new Date(props.transaction.timestamp * 1000));
