+++ src/i18n/en.po
@@ -372,12 +372,12 @@ msgstr ""
 msgid "Back to Bank Details"
 msgstr ""
 
-#: src/components/BtcTransactionListItem.vue:263
 #: src/components/modals/BtcTransactionModal.vue:505
 #: src/components/modals/TransactionModal.vue:562
 #: src/components/modals/UsdcTransactionModal.vue:421
-#: src/components/TransactionListItem.vue:242
-#: src/components/UsdcTransactionListItem.vue:260
+#: src/composables/useBtcTransactionInfo.ts:121
+#: src/composables/useTransactionInfo.ts:100
+#: src/composables/useUsdcTransactionInfo.ts:129
 msgid "Bank Account"
 msgstr ""
 
@@ -404,8 +404,8 @@ msgstr ""
 #: src/components/modals/BtcTransactionModal.vue:128
 #: src/components/modals/TransactionModal.vue:554
 #: src/components/modals/UsdcTransactionModal.vue:417
-#: src/components/TransactionListItem.vue:234
-#: src/components/UsdcTransactionListItem.vue:256
+#: src/composables/useTransactionInfo.ts:92
+#: src/composables/useUsdcTransactionInfo.ts:125
 msgid "Bitcoin"
 msgstr ""
 
@@ -525,15 +525,15 @@ msgstr ""
 msgid "Cancel Swap"
 msgstr ""
 
-#: src/components/BtcTransactionListItem.vue:248
 #: src/components/modals/BtcTransactionModal.vue:490
 #: src/components/modals/BtcTransactionModal.vue:6
 #: src/components/modals/TransactionModal.vue:5
 #: src/components/modals/TransactionModal.vue:549
 #: src/components/modals/UsdcTransactionModal.vue:406
 #: src/components/modals/UsdcTransactionModal.vue:6
-#: src/components/TransactionListItem.vue:229
-#: src/components/UsdcTransactionListItem.vue:245
+#: src/composables/useBtcTransactionInfo.ts:106
+#: src/composables/useTransactionInfo.ts:87
+#: src/composables/useUsdcTransactionInfo.ts:114
 msgid "Cancelled Swap"
 msgstr ""
 
@@ -543,7 +543,7 @@ msgstr ""
 
 #: src/components/modals/TransactionModal.vue:571
 #: src/components/TransactionList.vue:62
-#: src/components/TransactionListItem.vue:251
+#: src/composables/useTransactionInfo.ts:109
 msgid "Cashlink"
 msgstr ""
 
@@ -925,7 +925,6 @@ msgstr ""
 msgid "Exchange rate"
 msgstr ""
 
-#: src/components/BtcTransactionListItem.vue:45
 #: src/components/TransactionListItem.vue:39
 #: src/components/UsdcTransactionListItem.vue:37
 msgid "expired"
@@ -1003,7 +1002,7 @@ msgstr ""
 msgid "Fetching"
 msgstr ""
 
-#: src/components/BtcTransactionListItem.vue:66
+#: src/components/BtcTransactionListItem.vue:65
 #: src/components/modals/TransactionModal.vue:206
 #: src/components/TransactionListItem.vue:57
 #: src/components/UsdcTransactionListItem.vue:55
@@ -1129,26 +1128,26 @@ msgstr ""
 msgid "Historic value"
 msgstr ""
 
-#: src/components/BtcTransactionListItem.vue:208
+#: src/components/BtcTransactionListItem.vue:156
 #: src/components/modals/BtcTransactionModal.vue:454
 #: src/components/modals/TransactionModal.vue:488
-#: src/components/TransactionListItem.vue:186
-#: src/components/UsdcTransactionListItem.vue:190
+#: src/components/TransactionListItem.vue:156
+#: src/composables/useUsdcTransactionInfo.ts:171
 #: src/lib/DataFormatting.ts:21
 msgid "HTLC Creation"
 msgstr ""
 
-#: src/components/BtcTransactionListItem.vue:208
+#: src/components/BtcTransactionListItem.vue:156
 #: src/components/modals/BtcTransactionModal.vue:454
 #: src/components/modals/TransactionModal.vue:499
-#: src/components/TransactionListItem.vue:197
-#: src/components/UsdcTransactionListItem.vue:201
+#: src/components/TransactionListItem.vue:167
+#: src/composables/useUsdcTransactionInfo.ts:182
 msgid "HTLC Refund"
 msgstr ""
 
 #: src/components/modals/TransactionModal.vue:492
-#: src/components/TransactionListItem.vue:190
-#: src/components/UsdcTransactionListItem.vue:194
+#: src/components/TransactionListItem.vue:160
+#: src/composables/useUsdcTransactionInfo.ts:175
 msgid "HTLC Settlement"
 msgstr ""
 
@@ -1223,7 +1222,7 @@ msgstr ""
 msgid "Insufficient balance."
 msgstr ""
 
-#: src/components/BtcTransactionListItem.vue:46
+#: src/components/BtcTransactionListItem.vue:45
 #: src/components/TransactionListItem.vue:40
 #: src/components/UsdcTransactionListItem.vue:38
 msgid "invalid"
@@ -1880,9 +1879,9 @@ msgstr ""
 msgid "Sent {btc} BTC to {name}"
 msgstr ""
 
-#: src/components/BtcTransactionListItem.vue:193
-#: src/components/TransactionListItem.vue:169
-#: src/components/UsdcTransactionListItem.vue:173
+#: src/components/BtcTransactionListItem.vue:141
+#: src/components/TransactionListItem.vue:139
+#: src/composables/useUsdcTransactionInfo.ts:154
 msgid "Sent {fromAsset} – Received {toAsset}"
 msgstr ""
 
@@ -2044,7 +2043,6 @@ msgid "Super low fees, super fast payments"
 msgstr ""
 
 #. avoid displaying the proxy address until we know related peer address
-#: src/components/BtcTransactionListItem.vue:255
 #: src/components/layouts/Sidebar.vue:115
 #: src/components/modals/BtcTransactionModal.vue:497
 #: src/components/modals/BtcTransactionModal.vue:9
@@ -2052,8 +2050,9 @@ msgstr ""
 #: src/components/modals/TransactionModal.vue:7
 #: src/components/modals/UsdcTransactionModal.vue:413
 #: src/components/modals/UsdcTransactionModal.vue:9
-#: src/components/TransactionListItem.vue:225
-#: src/components/UsdcTransactionListItem.vue:252
+#: src/composables/useBtcTransactionInfo.ts:113
+#: src/composables/useTransactionInfo.ts:83
+#: src/composables/useUsdcTransactionInfo.ts:121
 msgid "Swap"
 msgstr ""
 
@@ -2253,12 +2252,12 @@ msgstr ""
 msgid "This might take up to {min} minutes."
 msgstr ""
 
-#: src/components/BtcTransactionList.vue:232
-#: src/components/BtcTransactionList.vue:254
-#: src/components/TransactionList.vue:247
-#: src/components/TransactionList.vue:269
-#: src/components/UsdcTransactionList.vue:197
-#: src/components/UsdcTransactionList.vue:219
+#: src/components/BtcTransactionList.vue:240
+#: src/components/BtcTransactionList.vue:262
+#: src/components/TransactionList.vue:252
+#: src/components/TransactionList.vue:274
+#: src/components/UsdcTransactionList.vue:202
+#: src/components/UsdcTransactionList.vue:224
 msgid "This month"
 msgstr ""
 
@@ -2351,7 +2350,7 @@ msgid "Tx time"
 msgstr ""
 
 #: src/components/modals/TransactionModal.vue:572
-#: src/components/TransactionListItem.vue:252
+#: src/composables/useTransactionInfo.ts:110
 msgid "Unclaimed Cashlink"
 msgstr ""
 
@@ -2393,13 +2392,13 @@ msgid "Update now"
 msgstr ""
 
 #: src/components/AddressList.vue:179
-#: src/components/BtcTransactionListItem.vue:259
 #: src/components/layouts/AccountOverview.vue:158
 #: src/components/layouts/AddressOverview.vue:84
 #: src/components/modals/BtcTransactionModal.vue:501
 #: src/components/modals/TransactionModal.vue:558
-#: src/components/TransactionListItem.vue:238
 #: src/components/UsdcAddressInfo.vue:12
+#: src/composables/useBtcTransactionInfo.ts:117
+#: src/composables/useTransactionInfo.ts:96
 msgid "USD Coin"
 msgstr ""
